# docs/swagger.yaml
definitions:
  main.chartNew:
    properties:
      chart:
        type: string
      readme:
        type: string
      repoName:
        type: string
      template:
        items:
          $ref: '#/definitions/main.file'
        type: array
      values:
        additionalProperties: true
        type: object
    type: object
  main.downFactor:
    properties:
      chartUrl:
        type: string
      repoUrl:
        type: string
    type: object
  main.file:
    properties:
      data:
        type: string
      name:
        type: string
    type: object
...
  main.respBody:
    properties:
      code:
        description: 0 or 1, 0 is ok, 1 is error
        type: integer
      data:
        type: object
      error:
        type: string
    type: object
info:
  contact:
    email: mika055@163.com
    name: polya
  description: This is a api proxy of helm.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Helm API Proxy
  version: 0.0.1
paths:
  /charts:
    get:
      description: 根据chart名称，获取chart的readme、values、chart、template信息
      parameters:
      - description: chart名称
        in: query
        name: chart
        required: true
        type: string
      - description: chart版本
        in: query
        name: version
        type: string
      - description: Enums(all, readme, values, chart、template)
        in: query
        name: info
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.respBody'
      summary: 获取chart详细信息
      tags:
      - Chart
  /charts/create:
    post:
      description: 新建一个chart并上传至镜像库
      parameters:
      - description: chart信息
        in: body
        name: newChart
        required: true
        schema:
          $ref: '#/definitions/main.chartNew'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.respBody'
      summary: 新建chart
      tags:
      - Chart
  /charts/export:
    post:
      description: 获取chart的下载地址
      parameters:
      - description: 带下载的chart路径信息
        in: body
        name: factor
        required: true
        schema:
          $ref: '#/definitions/main.downFactor'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.respBody'
      summary: 获取chart下载地址
      tags:
      - Chart
  /charts/template:
    post:
      description: 显示chart的k8s部署yaml，如果多个文件则合并到一个yaml一起展示出来
      parameters:
      - description: chart名称
        in: query
        name: chart
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.respBody'
      summary: 显示chart解析后的k8s部署yaml
      tags:
      - Chart
...
  /repos/charts:
    get:
      description: 在本地库中查找chart，如没有keyword则列出所有chart
      parameters:
      - description: 搜索关键字
        in: query
        name: keyword
        type: string
      - description: chart版本
        in: query
        name: version
        type: string
      - description: 如果true，查询出每个chart的所有版本；false，只列出每个chart最新版
        in: query
        name: versions
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.respBody'
      summary: 查找chart/列出本地库中的所有chart
      tags:
      - Repository
...
swagger: "2.0"